predix-python-linear-regression-analytic/requirements.txt
Flask
requests
flask-cors
websocket-client
scikit-learn
pandas
scipy
